webapp/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
qrcode==7.3.1
Pillow==8.4.0
Werkzeug==2.2.2
SQLAlchemy==1.3.23
python-dotenv==1.0.1